requirements.txt
torch
boto3

pytorch-lightning
datasets
ipdb
hydra-core
jsonlines
scipy
clearml
tqdm
torch-scatter -f https://data.pyg.org/whl/torch-1.11.0+cu115.html
torch-sparse -f https://data.pyg.org/whl/torch-1.11.0+cu115.html
torch-cluster -f https://data.pyg.org/whl/torch-1.11.0+cu115.html
torch-spline-conv -f https://data.pyg.org/whl/torch-1.11.0+cu115.html 
torch-geometric -f https://data.pyg.org/whl/torch-1.11.0+cu115.html